Conflicts:
	app/code/core/Mage/Webapi/view/adminhtml/layout.xml

+++ app/code/core/Mage/Webapi/controllers/Adminhtml/Webapi/UserController.php
@@ -69,18 +69,9 @@ class Mage_Webapi_Adminhtml_Webapi_UserController extends Mage_Backend_Controlle
              ->_title(Mage::helper('Mage_Webapi_Helper_Data')->__('API Users'));
 
         $userId = (int)$this->getRequest()->getParam('user_id');
-
-        /** @var $user Mage_Webapi_Model_Acl_User */
-        $user = Mage::getModel('Mage_Webapi_Model_Acl_User');
-
-        if ($userId) {
-            $user->load($userId);
-            if (!$user->getId()) {
-                $this->_getSession()->addError(
-                    Mage::helper('Mage_Webapi_Helper_Data')->__('This user no longer exists.'));
-                $this->_redirect('*/*/');
-                return;
-            }
+        $user = $this->_loadApiUser($userId);
+        if (!$user) {
+            return;
         }
 
         // Update title and breadcrumb record
@@ -101,11 +92,6 @@ class Mage_Webapi_Adminhtml_Webapi_UserController extends Mage_Backend_Controlle
         if ($editBlock) {
             $editBlock->setApiUser($user);
         }
-        /** @var $tabsBlock Mage_Webapi_Block_Adminhtml_User_Edit_Tabs */
-        $tabsBlock = $this->getLayout()->getBlock('webapi.user.edit.tabs');
-        if ($tabsBlock) {
-            $tabsBlock->setApiUser($user);
-        }
         $this->renderLayout();
     }
 
@@ -117,12 +103,8 @@ class Mage_Webapi_Adminhtml_Webapi_UserController extends Mage_Backend_Controlle
         $data = $this->getRequest()->getPost();
         if ($data) {
             $userId = (int)$this->getRequest()->getPost('user_id');
-            /** @var $user Mage_Webapi_Model_Acl_User */
-            $user = Mage::getModel('Mage_Webapi_Model_Acl_User')->load($userId);
-            if (!$user->getId() && $userId) {
-                $this->_getSession()->addError(
-                    Mage::helper('Mage_Webapi_Helper_Data')->__('This user no longer exists.'));
-                $this->_redirect('*/*/');
+            $user = $this->_loadApiUser($userId);
+            if (!$user) {
                 return;
             }
 
@@ -132,9 +114,11 @@ class Mage_Webapi_Adminhtml_Webapi_UserController extends Mage_Backend_Controlle
 
                 $this->_getSession()->addSuccess(
                     Mage::helper('Mage_Webapi_Helper_Data')->__('The user has been saved.'));
-                $this->_getSession()->setWebapiUserData(false);
-                $this->_redirect('*/*/edit', array('user_id' => $user->getId()));
-                return;
+                $this->_getSession()->setWebapiUserData(null);
+                if ($this->getRequest()->getParam('back')) {
+                    $this->_redirect('*/*/edit', array('user_id' => $user->getId()));
+                    return;
+                }
             } catch (Exception $e) {
                 $this->_getSession()->addError($e->getMessage());
                 $this->_getSession()->setWebapiUserData($data);
@@ -152,9 +136,11 @@ class Mage_Webapi_Adminhtml_Webapi_UserController extends Mage_Backend_Controlle
     {
         $userId = (int)$this->getRequest()->getParam('user_id');
         if ($userId) {
+            $user = $this->_loadApiUser($userId);
+            if (!$user) {
+                return;
+            }
             try {
-                /** @var $user Mage_Webapi_Model_Acl_User */
-                $user = Mage::getModel('Mage_Webapi_Model_Acl_User')->load($userId);
                 $user->delete();
 
                 $this->_getSession()->addSuccess(
@@ -173,23 +159,11 @@ class Mage_Webapi_Adminhtml_Webapi_UserController extends Mage_Backend_Controlle
     }
 
     /**
-     * Web API User Roles grid
+     * AJAX Web API Users grid
      */
-    public function rolesAction()
+    public function gridAction()
     {
-        $userId = (int)$this->getRequest()->getParam('user_id');
-        /** @var $user Mage_Webapi_Model_Acl_User */
-        $user = Mage::getModel('Mage_Webapi_Model_Acl_User');
-        if ($userId) {
-            $user->load($userId);
-        }
-
         $this->loadLayout();
-        /** @var $gridBlock Mage_Webapi_Block_Adminhtml_User_Edit_Tab_Roles */
-        $gridBlock = $this->getLayout()->getBlock('root');
-        if ($gridBlock) {
-            $gridBlock->setApiUser($user);
-        }
         $this->renderLayout();
     }
 
@@ -201,5 +175,24 @@ class Mage_Webapi_Adminhtml_Webapi_UserController extends Mage_Backend_Controlle
     protected function _isAllowed()
     {
         return Mage::getSingleton('Mage_Core_Model_Authorization')->isAllowed('Mage_Webapi::webapi_users');
+    }
+
+    /**
+     * Load Web API User
+     *
+     * @param $userId
+     * @return bool|Mage_Webapi_Model_Acl_User
+     */
+    protected function _loadApiUser($userId)
+    {
+        /** @var $user Mage_Webapi_Model_Acl_User */
+        $user = Mage::getModel('Mage_Webapi_Model_Acl_User')->load($userId);
+        if (!$user->getId() && $userId) {
+            $this->_getSession()->addError(
+                Mage::helper('Mage_Webapi_Helper_Data')->__('This user no longer exists.'));
+            $this->_redirect('*/*/');
+            return false;
+        }
+        return $user;
     }
 }
